requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
httpx
python-dotenv
aiofiles
